RPA development

+++ src/AluraSearcherRPA.Application/Interfaces/IHistoryService.cs
@@ -5,7 +5,7 @@ namespace AluraSearcherRPA.Application.Interfaces
     public interface IHistoryService
     {
         IEnumerable<HistoryResponseDTO> GetAllHistory();
-        HistoryResponseDTO GetHistory(int id);
+        HistoryResponseDTO GetHistory(long id);
         IEnumerable<HistoryResponseDTO> GetHistory(string searchedValue);
     }
 }
